+++ internal/case/backjob/updatetelegramaccountmessage/resolve.go
@@ -18,6 +18,7 @@ import (
 
 type Env interface {
 	Logger() logger.Logger
+	LogLevel() string
 	GetTelegramPublishSentAccountMessageContentHash(ctx context.Context, arg db.GetTelegramPublishSentAccountMessageContentHashParams) (string, error)
 	GetTelegramPublishSentAccountMessagePostType(ctx context.Context, arg db.GetTelegramPublishSentAccountMessagePostTypeParams) (string, error)
 	UpdateTelegramPublishSentAccountMessageContent(ctx context.Context, arg db.UpdateTelegramPublishSentAccountMessageContentParams) error